Add InAppWebView plugin for displaying frequently visited websites in Aspirant App.
We have integrated the InAppWebView plugin to enhance the user experience in the Aspirant App. We are now showing an in-app view of several useful websites that are frequently visited by the aspirants of the library, all on a single page. This feature allows users to access important online resources seamlessly within the app. The InAppWebView plugin plays a crucial role in enabling this functionality, providing a smooth and efficient browsing experience.